requirements.txt
selenium~=4.27.1
requests~=2.32.3
Pillow~=10.0.0